# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "edcurate-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/presets/": {
      "post": {
        "tags": [
          "presets"
        ],
        "summary": "Create Preset",
        "description": "Create a new classroom preset.",
        "operationId": "create_preset_api_presets__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "presets"
        ],
        "summary": "List Presets",
        "description": "List classroom presets for the current user (paginated).",
        "operationId": "list_presets_api_presets__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_PresetResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/presets/{preset_id}": {
      "get": {
        "tags": [
          "presets"
        ],
        "summary": "Get Preset",
        "description": "Get a single classroom preset by id.",
        "operationId": "get_preset_api_presets__preset_id__get",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "presets"
        ],
        "summary": "Update Preset",
        "description": "Full update of a classroom preset.",
        "operationId": "update_preset_api_presets__preset_id__put",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Preset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "presets"
        ],
        "summary": "Delete Preset",
        "description": "Delete a classroom preset. Returns 409 if it is the default.",
        "operationId": "delete_preset_api_presets__preset_id__delete",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/presets/{preset_id}/default": {
      "patch": {
        "tags": [
          "presets"
        ],
        "summary": "Set Default Preset",
        "description": "Set a classroom preset as the default (atomic swap).",
        "operationId": "set_default_preset_api_presets__preset_id__default_patch",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetDefaultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_PresetResponse_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PresetResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "PaginatedResponse[PresetResponse]"
      },
...
      "PresetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          },
          "curriculum_framework": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculum Framework"
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Subject"
          },
          "year_level": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Year Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Country"
          },
          "state_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "State Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "teaching_language": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Teaching Language",
            "default": "en"
          },
          "class_size": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Size"
          },
          "eal_d_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Eal D Students"
          },
          "reading_support_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Support Students"
          },
          "extension_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Students"
          },
          "student_interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Student Interests"
          },
          "language_backgrounds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Language Backgrounds"
          },
          "average_reading_level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Reading Level"
          },
          "source_weights": {
            "$ref": "#/components/schemas/SourceWeights"
          },
          "additional_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "subject",
          "year_level",
          "country"
        ],
        "title": "PresetCreate",
        "description": "Schema for creating a classroom preset."
      },
      "PresetDefaultResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          }
        },
        "type": "object",
        "required": [
          "id",
          "is_default"
        ],
        "title": "PresetDefaultResponse",
        "description": "Minimal response after setting a preset as default."
      },
      "PresetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "curriculum_framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculum Framework"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "year_level": {
            "type": "string",
            "title": "Year Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "state_region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "teaching_language": {
            "type": "string",
            "title": "Teaching Language"
          },
          "class_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Size"
          },
          "eal_d_students": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eal D Students"
          },
          "reading_support_students": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Support Students"
          },
          "extension_students": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Students"
          },
          "student_interests": {
            "items": {},
            "type": "array",
            "title": "Student Interests"
          },
          "language_backgrounds": {
            "items": {},
            "type": "array",
            "title": "Language Backgrounds"
          },
          "average_reading_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Reading Level"
          },
          "source_weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Weights"
          },
          "additional_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "is_default",
          "name",
          "curriculum_framework",
          "subject",
          "year_level",
          "topic",
          "country",
          "state_region",
          "city",
          "teaching_language",
          "class_size",
          "eal_d_students",
          "reading_support_students",
          "extension_students",
          "student_interests",
          "language_backgrounds",
          "average_reading_level",
          "source_weights",
          "additional_notes",
          "created_at",
          "updated_at"
        ],
        "title": "PresetResponse",
        "description": "Schema for preset response including id and timestamps."
      },
      "PresetUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          },
          "curriculum_framework": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculum Framework"
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Subject"
          },
          "year_level": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Year Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "country": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Country"
          },
          "state_region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "State Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "teaching_language": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Teaching Language",
            "default": "en"
          },
          "class_size": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Size"
          },
          "eal_d_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Eal D Students"
          },
          "reading_support_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Support Students"
          },
          "extension_students": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Students"
          },
          "student_interests": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Student Interests"
          },
          "language_backgrounds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Language Backgrounds"
          },
          "average_reading_level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Reading Level"
          },
          "source_weights": {
            "$ref": "#/components/schemas/SourceWeights"
          },
          "additional_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "subject",
          "year_level",
          "country"
        ],
        "title": "PresetUpdate",
        "description": "Schema for updating a classroom preset (full replacement)."
      },
...
    }
  }
}